reverting neo/Prey/game_anim.cpp (f940bfe), adding aarch64/arm64 to idEventArg guard in neo/game/gamesys/Class.h (addresses win x64 breakage seen in #26, raspbian/aarch64 in #27)

+++ neo/Prey/game_anim.cpp
@@ -151,7 +151,7 @@ bool hhAnim::CallFrameCommandsExtra( const frameCommand_t &command, idEntity *en
 	switch( command.type ) {
 		case FC_EVENT_ARGS: {
 			if ( command.function && command.parmList && command.function->eventdef ) {
-				ent->ProcessEvent( command.function->eventdef, (int)(intptr_t)command.parmList );
+				ent->ProcessEvent( command.function->eventdef, (intptr_t)command.parmList );
 			}	
 			return( true );
 		}
